requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
pydantic==2.6.3

# Supabase client for Python.  This library allows you to interact with
# Supabase's REST API and authentication services.
supabase-py==2.1.1

# Optional: environment variable loading
python-dotenv==1.0.1

# HTTP client used by the reverse proxy endpoint
httpx==0.25.2